Struct dslab_mr::dag::Dag

pub struct Dag { /* private fields */ }

Represents DAG of stages.

Implementations

impl Dag

pub fn new() -> Self

Creates new empty DAG.

pub fn add_initial_data(&mut self, stage_id: usize, initial_data: u64)

Adds initial data for a stage.

pub fn initial_data(&self) -> &BTreeMap<usize, u64>

Returns map of initial data sizes for each stage.

pub fn add_stage( &mut self, tasks: Vec<Box<dyn Task>>, upload_result_to_dfs: bool ) -> usize

Add new stage.

pub fn add_connection( &mut self, from: usize, to: usize, shuffle: Option<Box<dyn Shuffle>> ) -> usize

Add new connection between stages.

pub fn mark_started(&mut self, stage_id: usize)

Mark stage as started.

pub fn mark_completed(&mut self, stage_id: usize)

Mark stage as completed.

pub fn ready_stages(&self) -> &BTreeSet<usize>

Returns set of all stage ready to be started.

pub fn running_stages(&self) -> &BTreeSet<usize>

Returns set of all running stage.

pub fn completed_stages(&self) -> &BTreeSet<usize>

Returns set of all completed stage.

pub fn stages(&self) -> &Vec<Stage>

Returns vector of all stages.

pub fn stage(&self, stage_id: usize) -> &Stage

Returns stage by id.

pub fn connection(&self, connection_id: usize) -> &Connection

Returns connection by id.

pub fn stage_dependencies(&self, stage_id: usize) -> &Vec<Connection>

Returns stage dependencies.

pub fn outgoing_connections(&self, stage_id: usize) -> &Vec<usize>

Returns stage outdoing connections.

impl Dag

pub fn from_yaml<P: AsRef<Path>>(file: P) -> Self

Read Dag from YAML file. Panic on error.